feat: hide extra amenities / vertically align hotel cards
We can improve the vertical rhythm of the page by removing the negative space
produced by the irregular number of amenities per hotel.

We do this by only showing the first X amenities. We choose a number small
enough that most hotels have at least that many and we can tolerate the few
outliers.

We could also get rid of the negative space with `grid-auto-flow: dense`. That
has two trade-offs. One, listing all of the amenities is noisy, so we'd rather
list the more popular ones first (not implemented here yet). Two, that property
would "pack" the grid items, which would result in a more organic flowing
layout. While pretty for a photo gallery, it interferes with horizontal scanning
patterns.